bot/requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.1
loguru>=0.7.2
pytest>=8.0.0
pytest-asyncio>=0.23.0
